Fix broken stories after moving LessonLayout
