+++ stories/components/LessonCard.stories.tsx
@@ -67,8 +67,8 @@ export const Basic: React.FC = () => (
     title={`HTML, CSS, JavaScript`}
     challengeCount={7}
     description={`Create basic Front-End Mini-Projects that demonstrate User Interface logic and understanding of Web Development.`}
-    reviewUrl="https://c0d3.com/teacher/5"
-    challengesUrl="https://c0d3.com/student/24"
+    reviewUrl="/review/js0"
+    challengesUrl="/curriculum/js0"
     docUrl="https://www.notion.so/JS-4-Front-End-Engineering-c59fbdd58dcc4214956f7856e0892b52"
   />
 )
@@ -81,8 +81,8 @@ export const withInProgress: React.FC = () => (
     challengeCount={7}
     description={`Create basic Front-End Mini-Projects that demonstrate User Interface logic and understanding of Web Development.`}
     currentState="inProgress"
-    reviewUrl="https://c0d3.com/teacher/5"
-    challengesUrl="https://c0d3.com/student/24"
+    reviewUrl="/review/js0"
+    challengesUrl="/curriculum/js0"
     docUrl="https://www.notion.so/JS-4-Front-End-Engineering-c59fbdd58dcc4214956f7856e0892b52"
   />
 )
@@ -96,8 +96,8 @@ export const withCompleted: React.FC = () => (
       challengeCount={7}
       description={`Create basic Front-End Mini-Projects that demonstrate User Interface logic and understanding of Web Development.`}
       currentState="completed"
-      reviewUrl="https://c0d3.com/teacher/5"
-      challengesUrl="https://c0d3.com/student/24"
+      reviewUrl="/review/js0"
+      challengesUrl="/curriculum/js0"
       docUrl="https://www.notion.so/JS-4-Front-End-Engineering-c59fbdd58dcc4214956f7856e0892b52"
     />
   </MockedProvider>
