+++ stories/components/mdx/BlockQuote.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import BlockQuote from '../../../components/mdx/BlockQuote'
-import Layout from '../../../components/mdx/LessonLayout'
 import { MockedProvider } from '@apollo/client/testing'
 import GET_APP from '../../../graphql/queries/getApp'
 import dummyLessonData from '../../../__dummy__/lessonData'
@@ -26,19 +25,16 @@ const mocks = [
 export const Quote: React.FC = () => {
   return withTestRouter(
     <MockedProvider mocks={mocks} addTypename={false}>
-      <Layout>
-        <BlockQuote>
-          <p>
-            Ipsum molestiae natus adipisci modi eligendi? Debitis amet quae unde
-            commodi aspernatur enim, consectetur. Cumque deleniti temporibus
-            ipsam atque a dolores quisquam quisquam adipisci possimus
-            laboriosam. Quibusdam facilis doloribus debitis! Sit quasi quod
-            accusamus eos quod. Ab quos consequuntur eaque quo rem! Mollitia
-            reiciendis porro quo magni incidunt dolore amet atque facilis ipsum
-            deleniti rem!
-          </p>
-        </BlockQuote>
-      </Layout>
+      <BlockQuote>
+        <p>
+          Ipsum molestiae natus adipisci modi eligendi? Debitis amet quae unde
+          commodi aspernatur enim, consectetur. Cumque deleniti temporibus ipsam
+          atque a dolores quisquam quisquam adipisci possimus laboriosam.
+          Quibusdam facilis doloribus debitis! Sit quasi quod accusamus eos
+          quod. Ab quos consequuntur eaque quo rem! Mollitia reiciendis porro
+          quo magni incidunt dolore amet atque facilis ipsum deleniti rem!
+        </p>
+      </BlockQuote>
     </MockedProvider>
   )
 }
